move helper functions to _utils.py
